+++ .github/workflows/ci.yml
@@ -5,7 +5,7 @@ on:
     branches:
       - master
 env:
-  GOPRIVATE: github.com/sylabs/compute-cli
+  GOPRIVATE: github.com/sylabs/fuzzctl
 
 jobs:
   build-and-unit-test:
