+++ internal/pkg/client/round_tripper.go
@@ -1,6 +1,6 @@
 // Copyright (c) 2020, Sylabs, Inc. All rights reserved.
 
-package compute
+package client
 
 import "net/http"
